% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{higher_order_terms}
\alias{higher_order_terms}
\title{Higher order terms for optimal transport sensitivity indices}
\usage{
higher_order_terms(ot_result, wb_result)
}
\arguments{
\item{ot_result}{An object returned by \code{\link[=ot_indices]{ot_indices()}}.}

\item{wb_result}{An object returned by \code{\link[=ot_indices_wb]{ot_indices_wb()}}.}
}
\value{
An object of class \code{gsaot_indices} containing the higher order terms
of the Wasserstein-Bures decomposition.
}
\description{
Compute the higher order terms as the difference between the output of
\code{\link[=ot_indices]{ot_indices()}} and the output of \code{\link[=ot_indices_wb]{ot_indices_wb()}} computed on the same
sample.
}
\details{
The helper only computes the point estimate difference between two
already computed results. The function does not check that the ground cost
used for the \code{ot_result} object is the squared Euclidean one (default). The
user should therefore pay attention to this aspect when using the function.
}
\examples{
dat <- gaussian_fun(1000)
ot_result <- ot_indices(dat$x, dat$y, 10)
wb_result <- ot_indices_wb(dat$x, dat$y, 10)
higher_order_terms(ot_result, wb_result)

}
