% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ot_indices_bw.R
\name{ot_indices_bw}
\alias{ot_indices_bw}
\title{Evaluate Wasserstein-Bures approximation of the Optimal Transport solution}
\usage{
ot_indices_bw(x, y, M, ext_out = FALSE)
}
\arguments{
\item{x}{a data.frame containing the input(s) values}

\item{y}{a data.frame containing the outputs values}

\item{M}{a scalar representing the number of partitions for continuous inputs}

\item{ext_out}{logical indicating if the function should return the inner statistics and the partitions}
}
\value{
A sensitivity index between 0 and 1 for each of the columns in x
}
\description{
Evaluate Wasserstein-Bures approximation of the Optimal Transport solution
}
\examples{
N <- 10000

mx <- c(1, 1, 1)
Sigmax <- matrix(data = c(1, 0.5, 0.5, 0.5, 1, 0.5, 0.5, 0.5, 1), nrow = 3)

x1 <- rnorm(N)
x2 <- rnorm(N)
x3 <- rnorm(N)

x <- cbind(x1, x2, x3)
x <- mx + x \%*\% chol(Sigmax)

A <- matrix(data = c(4, -2, 1, 2, 5, -1), nrow = 2, byrow = TRUE)
y <- t(A \%*\% t(x))

x <- data.frame(x)
y <- y

ot_indices_bw(x, y, 100)
}
