% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ot_indices_entropic.R
\name{ot_indices_entropic}
\alias{ot_indices_entropic}
\title{Title}
\usage{
ot_indices_entropic(
  x,
  y,
  M,
  eps,
  d = 2,
  cost = "euclidean",
  num_iterations = 1e+06
)
}
\arguments{
\item{x}{a data.frame containing the input(s) values}

\item{y}{a data.frame containing the outputs values}

\item{M}{a scalar representing the number of partitions for continuous inputs}

\item{eps}{a double representing the coefficient of the entropic regularization. If negative, the cost matrix is normalized.}

\item{d}{exponential of the cost}

\item{cost}{type of cost. The types are defined by the function dist of the stats package}

\item{num_iterations}{maximum number of iterations of the Sinkhorn algorithm}
}
\value{
A sensitivity index between 0 and 1 for each of the columns in x
}
\description{
Title
}
\examples{
N <- 1000

mx <- c(1, 1, 1)
Sigmax <- matrix(data = c(1, 0.5, 0.5, 0.5, 1, 0.5, 0.5, 0.5, 1), nrow = 3)

x1 <- rnorm(N)
x2 <- rnorm(N)
x3 <- rnorm(N)

x <- cbind(x1, x2, x3)
x <- mx + x \%*\% chol(Sigmax)

A <- matrix(data = c(4, -2, 1, 2, 5, -1), nrow = 2, byrow = TRUE)
y <- t(A \%*\% t(x))

x <- data.frame(x)
y <- data.frame(y)

ot_indices_entropic(x, y, 25, -0.01)
}
