% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsaot_indices.R
\name{summary.gsaot_indices}
\alias{summary.gsaot_indices}
\title{Summary method for \code{gsaot_indices} objects}
\usage{
\method{summary}{gsaot_indices}(object, digits = 3, ranking = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{"gsaot_indices"}.}

\item{digits}{(default: \code{3}) Number of significant digits to print for
numeric values.}

\item{ranking}{An integer with absolute value less or equal than the number
of inputs. If positive, select the first \code{ranking} inputs per importance.}

\item{...}{Further arguments (currently ignored).}
}
\value{
(Invisibly) a named \code{list} containing the main elements
summarised on screen.
}
\description{
Summary method for \code{gsaot_indices} objects
}
