% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ot_indices.R
\name{ot_indices}
\alias{ot_indices}
\title{Title}
\usage{
ot_indices(
  x,
  y,
  M,
  solver = "sinkhorn",
  solver_optns = NULL,
  scaling = TRUE,
  extended_out = FALSE
)
}
\arguments{
\item{x}{a data.frame containing the input(s) values}

\item{y}{a matrix containing the outputs values}

\item{M}{a scalar representing the number of partitions for continuous inputs}

\item{solver}{maximum number of iterations of the Sinkhorn algorithm}

\item{solver_optns}{algorithm to solve the Optimal Transport problem}

\item{scaling}{Lore ipsum}

\item{extended_out}{a double representing the coefficient of the entropic regularization. If negative, the cost matrix is normalized.}
}
\value{
A sensitivity index between 0 and 1 for each column in x
}
\description{
Title
}
\examples{
N <- 1000

mx <- c(1, 1, 1)
Sigmax <- matrix(data = c(1, 0.5, 0.5, 0.5, 1, 0.5, 0.5, 0.5, 1), nrow = 3)

x1 <- rnorm(N)
x2 <- rnorm(N)
x3 <- rnorm(N)

x <- cbind(x1, x2, x3)
x <- mx + x \%*\% chol(Sigmax)

A <- matrix(data = c(4, -2, 1, 2, 5, -1), nrow = 2, byrow = TRUE)
y <- t(A \%*\% t(x))

x <- data.frame(x)


}
