% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ot_indices_1d.R
\name{ot_indices_1d}
\alias{ot_indices_1d}
\title{Evaluate Optimal Transport indices on one dimensional outputs}
\usage{
ot_indices_1d(x, y, M, extended_out = FALSE)
}
\arguments{
\item{x}{A data.frame containing the input(s) values. The values can be
numeric, factors or strings.}

\item{y}{an array containing the output values}

\item{M}{A scalar representing the number of partitions for continuous
inputs.}

\item{extended_out}{(default \code{FALSE}) Logical indicating if the function
should return the inner statistics and the partitions.}
}
\value{
An Optimal Transport sensitivity index between 0 and 1 for each of
the columns in x
}
\description{
Evaluate Optimal Transport indices on one dimensional outputs
}
\examples{
x <- rnorm(1000)
y <- 10 * x
ot_indices_1d(data.frame(x), y, 30)
}
\seealso{
\code{\link[=ot_indices]{ot_indices()}}, \code{\link[=ot_indices_wb]{ot_indices_wb()}}
}
