% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ot_indices_1d.R
\name{ot_indices_1d}
\alias{ot_indices_1d}
\title{Evaluate Optimal Transport indices on one dimensional outputs}
\usage{
ot_indices_1d(x, y, M, grid = 0.001)
}
\arguments{
\item{x}{A data.frame containing the input(s) values. The values can be
numeric, factors or strings.}

\item{y}{An array containing the output values}

\item{M}{A scalar representing the number of partitions for continuous
inputs.}

\item{grid}{The spacing of the grid over which the integral of \eqn{\lVert F_Y - F_{Y|X_i \in \mathcal{X}_i} \rVert}}
}
\value{
An Optimal Transport sensitivity index between 0 and 1 for each of
the columns in x
}
\description{
Evaluate Optimal Transport indices on one dimensional outputs
}
\examples{
x <- rnorm(1000)
y <- 10 * x
ot_indices_1d(data.frame(x), y, 30)
}
\seealso{
\code{\link[=ot_indices]{ot_indices()}}, \code{\link[=ot_indices_wb]{ot_indices_wb()}}
}
