% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ot_indices_smap.R
\name{ot_indices_smap}
\alias{ot_indices_smap}
\title{Evaluate one-dimensional Optimal Transport indices on multivariate outputs}
\usage{
ot_indices_smap(x, y, M)
}
\arguments{
\item{x}{a data.frame containing the input(s) values}

\item{y}{a matrix containing the output(s) values}

\item{M}{a scalar representing the number of partitions for continuous inputs}
}
\value{
An Optimal Transport sensitivity index between 0 and 1 for each of the columns in x
}
\description{
Evaluate one-dimensional Optimal Transport indices on multivariate outputs
}
\examples{
x <- rnorm(1000)
y <- 10 * x
ot_indices_1d(data.frame(x), y, 30)
}
