% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ot_indices_bw.R
\name{ot_indices_wb}
\alias{ot_indices_wb}
\title{Evaluate Wasserstein-Bures approximation of the Optimal Transport solution}
\usage{
ot_indices_wb(
  x,
  y,
  M,
  boot = FALSE,
  R = NULL,
  parallel = "no",
  ncpus = 1,
  conf = 0.95,
  type = "norm"
)
}
\arguments{
\item{x}{A data.frame containing the input(s) values. The values can be
numeric, factors or strings.}

\item{y}{A matrix containing the output values. Each column represents a
different output variable, and each row represents a different observation.
Only numeric values are allowed.}

\item{M}{A scalar representing the number of partitions for continuous
inputs.}
}
\value{
An object containing:
\itemize{
\item \code{indices}: sensitivity indices between 0 and 1 for each column in x, indicating the influence of each input variable on the output variables.
\item \code{IS}: values of the inner statistics for the partitions defined by \code{partitions}.
\item \code{partitions}: the partitions built to calculate the sensitivity indices.
}
}
\description{
Evaluate Wasserstein-Bures approximation of the Optimal Transport solution
}
\examples{
N <- 1000

mx <- c(1, 1, 1)
Sigmax <- matrix(data = c(1, 0.5, 0.5, 0.5, 1, 0.5, 0.5, 0.5, 1), nrow = 3)

x1 <- rnorm(N)
x2 <- rnorm(N)
x3 <- rnorm(N)

x <- cbind(x1, x2, x3)
x <- mx + x \%*\% chol(Sigmax)

A <- matrix(data = c(4, -2, 1, 2, 5, -1), nrow = 2, byrow = TRUE)
y <- t(A \%*\% t(x))

x <- data.frame(x)
y <- y

ot_indices_wb(x, y, 100)
}
\seealso{
\code{\link[=ot_indices]{ot_indices()}}, \code{\link[=ot_indices_1d]{ot_indices_1d()}}
}
